requirements.txt
streamlit
pandas
numpy
openpyxl
google-api-python-client
google-auth
google-auth-httplib2
python-dotenv
bcrypt==4.1.2
reportlab==4.0.9
